fix(Format) fix rebase artifacts and new places where uint32_t prepended size appears

+++ nes-nautilus/include/Arena.hpp
@@ -56,7 +56,7 @@ struct ArenaRef
     /// Allocates memory from the arena. If the available space for the pointer is smaller than the required size, we allocate a new buffer from the arena.
     nautilus::val<int8_t*> allocateMemory(const nautilus::val<size_t>& sizeInBytes);
 
-    VariableSizedData allocateVariableSizedData(const nautilus::val<uint32_t>& sizeInBytes);
+    VariableSizedData allocateVariableSizedData(const nautilus::val<uint64_t>& sizeInBytes);
 
     [[nodiscard]] nautilus::val<Arena*> getArena() const;
